requirements.txt
langchain
langchain-openai
python-dotenv
requests
serpapi
google-search-results
markdown
xhtml2pdf
reportlab